+++ sentry-debug-images/src/unix.rs
@@ -4,7 +4,7 @@ use sentry_core::protocol::debugid::DebugId;
 use sentry_core::protocol::{DebugImage, SymbolicDebugImage};
 use sentry_core::types::Uuid;
 
-use findshlibs::{Segment, SharedLibrary, SharedLibraryId, TargetSharedLibrary, TARGET_SUPPORTED};
+use findshlibs::{SharedLibrary, SharedLibraryId, TargetSharedLibrary, TARGET_SUPPORTED};
 
 const UUID_SIZE: usize = 16;
 
@@ -50,29 +50,8 @@ pub fn debug_images() -> Vec<DebugImage> {
             None => return,
         };
 
-        let mut lowest_addr = !0;
-        let mut lowest_vmaddr = !0;
-        let mut highest_addr = 0;
-
-        for seg in shlib.segments() {
-            if !seg.is_code() {
-                continue;
-            }
-            let svma: u64 = seg.stated_virtual_memory_address().0 as u64;
-            let avma: u64 = seg.actual_virtual_memory_address(shlib).0 as u64;
-            if lowest_addr > avma {
-                lowest_addr = avma;
-            }
-            if highest_addr < avma {
-                highest_addr = avma;
-            }
-            if lowest_vmaddr > svma {
-                lowest_vmaddr = svma;
-            }
-        }
-
         let mut name = shlib.name().to_string_lossy().to_string();
-        if name == "" {
+        if name.is_empty() {
             name = env::current_exe()
                 .map(|x| x.display().to_string())
                 .unwrap_or_else(|_| "<main>".to_string());
@@ -82,9 +61,9 @@ pub fn debug_images() -> Vec<DebugImage> {
             SymbolicDebugImage {
                 name,
                 arch: None,
-                image_addr: lowest_addr.into(),
-                image_size: highest_addr - lowest_addr,
-                image_vmaddr: lowest_vmaddr.into(),
+                image_addr: shlib.actual_load_addr().0.into(),
+                image_size: shlib.len() as u64,
+                image_vmaddr: shlib.stated_load_addr().0.into(),
                 id: debug_id,
             }
             .into(),